Cambiar int por long para evitar problemas

+++ Main.java
@@ -4,17 +4,17 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int n = sc.nextLong();
+        long n = sc.nextLong();
 
-        int sumaT = n * (n + 1) / 2; //Algoritmo de suma de Gauss, referencia: https://nkhilv.medium.com/using-gauss-summation-to-add-numbers-quicker-2ab61d01eff1 
+        long sumaT = n * (n + 1) / 2; //Algoritmo de suma de Gauss, referencia: https://nkhilv.medium.com/using-gauss-summation-to-add-numbers-quicker-2ab61d01eff1 
 
-        int sumaR = 0;
+        long sumaR = 0;
 
-        for (int i = 0; i <= n - 1; i++) { 
+        for (long i = 0; i <= n - 1; i++) { 
             sumaReal += sc.nextLong();
         }
 
-        int numeroFaltante = sumaTeorica - sumaReal;
+        long numeroFaltante = sumaTeorica - sumaReal;
         
         System.out.println(numeroFaltante);
     }
